requirements.txt
discord.py==1.5.1
pyyaml==5.3
PyNaCl==1.4.0
